requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SSLify==0.1.5
frozenlist==1.3.3
google-api-core==2.11.0
googleapis-common-protos==1.58.0
google-auth==2.16.2
google-cloud-videointelligence==2.11.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.4
pip==23.0
protobuf==4.22.0
proto-plus==1.22.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
setuptools==67.1.0
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
